feat：打开新的页面展示关系图谱，并且添加节点和连线配色
