+++ frontend/package.json
@@ -15,10 +15,12 @@
     "@dnd-kit/sortable": "^9.0.0",
     "@dnd-kit/utilities": "^3.2.2",
     "@types/canvas-confetti": "^1.9.0",
+    "@types/dagre": "^0.7.54",
     "@xyflow/react": "^12.10.1",
     "antd": "^5.27.6",
     "axios": "^1.12.2",
     "canvas-confetti": "^1.9.4",
+    "dagre": "^0.8.5",
     "dayjs": "^1.11.13",
     "react": "^18.3.1",
     "react-diff-viewer-continued": "^3.4.0",
